chore(app): refactor models add start script
